% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelnh_LINUX.R
\name{parallelnh_LINUX}
\alias{parallelnh_LINUX}
\alias{paralllelnh_LINUX}
\title{Parallel NewHybrids - LINUX version}
\usage{
paralllelnh_LINUX(folder.data, where.NH, burnin, sweeps)
}
\arguments{
\item{folder.data}{A file path to the folder in which the NewHybrids formatted files to be analyzed reside}

\item{where.NH}{A file path to the NewHybrids installation folder}

\item{burnin}{An integer specifying how many burnin steps NewHybrids is to run}

\item{sweeps}{An integer specifying how many sweep steps NewHybrids is to run}
}
\description{
\code{paralllelnh_OSX} is the LINUX (Ubuntu) version of the function thatallows NewHybrids (Anderson XXXX) to be run in parallel. It does so by creating a vector of file names to be run, that are held within a single folder. paralleleNH_LINUX then assigns a job to each of the cores available in the computer. As each task finishes, parallelnh_LINUX assigns a new analysis to the idle core.
}

