% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelnh_OSX.R
\name{parallelnh_OSX}
\alias{parallelnh_OSX}
\title{Parallel NewHybrids - OS X version}
\usage{
parallelnh_OSX(folder.data, where.NH, burnin, sweeps)
}
\arguments{
\item{folder.data}{A file path to the folder in which the NewHybrids formatted files to be analyzed, and their associated individual file, reside. Note the path provided must end in "/"}

\item{where.NH}{A file path to the NewHybrids installation folder. Note this folder must be named "newhybrids", and the path provided must end in "/"}

\item{burnin}{An integer specifying how many burnin steps NewHybrids is to run}

\item{sweeps}{An integer specifying how many sweep steps NewHybrids is to run}
}
\description{
\code{paralllelnh_OSX} allows NewHybrids (Anderson XXXX) to be run in parallel. It does so by creating a vector of file names to be run, that are held within a single folder. paralleleNH_OSX then assigns a job to each of the cores available in the computer. As each task finishes, parallelnh_OSX assigns a new analysis to the idle core.
}

