% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimPops_S1R1_NH_individuals}
\alias{SimPops_S1R1_NH_individuals}
\title{SimPops_S1R1_NH_individuals}
\format{The individual file associated with SimPops_S1R1_NH. A single
copy of this file should be saved to the same folder in which SimPops_S1R1_NH
is saved. The filename must end in "individuals.txt".}
\source{
bla bla bla \url{http://www.blablabla.com/}
}
\usage{
SimPops_S1R1_NH_individuals
}
\description{
SimPops_S1R1_NH_individuals
}
\keyword{datasets}

